# Fix performance degradation while host functions call