--- a/examples/host-function-library/library/json-parser/export/library_plugin.pb.go
+++ b/examples/host-function-library/library/json-parser/export/library_plugin.pb.go
@@ -32,6 +32,7 @@ func (h parserLibrary) ParseJson(ctx context.Context, request *ParseJsonRequest)
 	}
 	ptr, size := wasm.ByteToPtr(buf)
 	ptrSize := _parse_json(ptr, size)
+	wasm.FreePtr(ptr)
 
 	ptr = uint32(ptrSize >> 32)
 	size = uint32(ptrSize)