*Issue #, if available:*
This is probably one of the option to fix performance issue  

*Description of changes:*
Hosts functions performance tests before fix
```bash
/u01/git/wasm/go-plugin/tests/host-functions
go test -run='^$' -bench '^Benchmark' ./... -count=6 -timeout 99999s
name              time/op
HostFunctions-12  1.76ms ±12%

name              alloc/op
HostFunctions-12  5.95kB ± 6%

name              allocs/op
HostFunctions-12    25.0 ± 0%
```

Hosts functions performance tests after fix
```bash
/u01/git/wasm/go-plugin/tests/host-functions
go test -run='^$' -bench '^Benchmark' ./... -count=6
HostFunctions-12  13.5µs ±12%

name              alloc/op
HostFunctions-12  6.27kB ± 1%

name              allocs/op
HostFunctions-12    25.0 ± 0%
```

Hosts functions performance tests for `v0.7.0`
```bash
/u01/git/wasm/go-plugin/tests/host-functions
go test -run='^$' -bench '^Benchmark' ./... -count=6
HostFunctions-12  10.6µs ±11%

name              alloc/op
HostFunctions-12  5.99kB ± 2%

name              allocs/op
HostFunctions-12    24.0 ± 0%
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

cc: @codefromthecrypt , @lburgazzoli