@codefromthecrypt 2023-04-20
ps you can use https://pkg.go.dev/golang.org/x/perf/cmd/benchstat to format the before after so it is easier to read.