@codefromthecrypt 2023-04-20
hmm I think there is a problem somewhere still, I ran offline just to see before/after and I see a lot of errors.

``` bash
$ go test -run='^$' -bench '^Benchmark.*' ./... -count 6
?   	github.com/knqyf263/go-plugin/cmd/protoc-gen-go-plugin	[no test files]
?   	github.com/knqyf263/go-plugin/encoding/defval	[no test files]
?   	github.com/knqyf263/go-plugin/encoding/tag	[no test files]
PASS
ok  	github.com/knqyf263/go-plugin/examples/helloworld	0.107s
?   	github.com/knqyf263/go-plugin/examples/helloworld/greeting	[no test files]
PASS
ok  	github.com/knqyf263/go-plugin/examples/host-function-library	0.072s
?   	github.com/knqyf263/go-plugin/examples/host-function-library/library/json-parser/export	[no test files]
?   	github.com/knqyf263/go-plugin/examples/host-function-library/library/json-parser/impl	[no test files]
?   	github.com/knqyf263/go-plugin/examples/host-function-library/proto	[no test files]
PASS
ok  	github.com/knqyf263/go-plugin/examples/host-functions	0.083s
?   	github.com/knqyf263/go-plugin/examples/host-functions/greeting	[no test files]
PASS
ok  	github.com/knqyf263/go-plugin/examples/known-types	0.079s
?   	github.com/knqyf263/go-plugin/examples/known-types/known	[no test files]
PASS
ok  	github.com/knqyf263/go-plugin/examples/wasi	0.080s
?   	github.com/knqyf263/go-plugin/examples/wasi/cat	[no test files]
?   	github.com/knqyf263/go-plugin/gen	[no test files]
?   	github.com/knqyf263/go-plugin/genid	[no test files]
?   	github.com/knqyf263/go-plugin/tests	[no test files]
goos: darwin
goarch: arm64
pkg: github.com/knqyf263/go-plugin/tests/fields
BenchmarkFields-12    	--- FAIL: BenchmarkFields-12
    bench_test.go:51: wasm error: out of bounds memory access
        wasm stack trace:
        	.runtime.hashmapSet(i32,i32,i32,i32)
        	.field_test_test(i32,i32) i64
BenchmarkFields-12    	--- FAIL: BenchmarkFields-12
    bench_test.go:51: wasm error: out of bounds memory access
        wasm stack trace:
        	.runtime.hashmapSet(i32,i32,i32,i32)
        	.field_test_test(i32,i32) i64
BenchmarkFields-12    	--- FAIL: BenchmarkFields-12
    bench_test.go:51: wasm error: out of bounds memory access
        wasm stack trace:
        	.runtime.hashmapSet(i32,i32,i32,i32)
        	.field_test_test(i32,i32) i64
BenchmarkFields-12    	^Csignal: interrupt
FAIL	github.com/knqyf263/go-plugin/tests/fields	0.468s
--snip--
panic: free: invalid pointer
--snip--
```